+++ package.json
@@ -132,6 +132,7 @@
     "rss": "^1.2.2",
     "screenfull": "^4.2.1",
     "simple-statistics": "^3.0.0",
+    "slug": "^8.2.2",
     "slugg": "^1.2.1",
     "speakingurl": "^14.0.1",
     "styled-components": "3.2.6",
